final year project/requirements.txt
sounddevice
numpy
scipy
noisereduce
pydub
sarvamai
python-dotenv
groq
asyncio
torch
onnxruntime
sentence-transformers
torchaudio<2.6.0
soundfile
streamlit
streamlit-mic-recorder
fastapi
uvicorn
python-multipart
httpx
